requirements.txt
alembic==1.13.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Jinja2==3.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
WTForms==3.1.1
Werkzeug==2.3.7
